Calling the method might be a good idea
